+++ internal/middleware/api_key_access.go
@@ -2,12 +2,15 @@ package middleware
 
 import (
 	"context"
+	stderrors "errors"
 	"net/http"
 	"strings"
 
 	"github.com/Tencent/WeKnora/internal/types"
 	"github.com/gin-gonic/gin"
 )
+
+var errTenantAPIKeyScopeForbidden = stderrors.New("tenant api key scope forbidden")
 
 // isTenantAPIKeyManagementPath reports tenant API-key lifecycle endpoints that
 // no scoped key may call, regardless of role.
@@ -19,6 +22,19 @@ func isTenantAPIKeyManagementPath(path string) bool {
 	}
 	// Legacy reset endpoint: POST /tenants/:id/api-key (singular).
 	return strings.HasSuffix(path, "/api-key")
+}
+
+// rejectTenantAPIKeyManagementPath is a defense-in-depth guard in auth: API keys
+// must never reach key-management routes even if a route forgets APIKeyDeny().
+// Role and KB checks are enforced by per-route APIKey* guards and types helpers.
+func rejectTenantAPIKeyManagementPath(ctx context.Context, path string) error {
+	if _, ok := types.TenantAPIKeyScopeFromContext(ctx); !ok {
+		return nil
+	}
+	if isTenantAPIKeyManagementPath(path) {
+		return errTenantAPIKeyScopeForbidden
+	}
+	return nil
 }
 
 // RequireAPIKeyDeny rejects all X-API-Key callers. JWT sessions pass through.
@@ -80,4 +96,13 @@ func requireTenantAPIKeyMinRole(ctx context.Context, minRole types.TenantRole) e
 		return errTenantAPIKeyScopeForbidden
 	}
 	return nil
+}
+
+func isSafeHTTPMethod(method string) bool {
+	switch method {
+	case http.MethodGet, http.MethodHead, http.MethodOptions:
+		return true
+	default:
+		return false
+	}
 }
